Restructure project layout with PostgreSQL setup and modularized models/schemas
