+++ backend/app/core/config.py
@@ -1,0 +1,16 @@
+import os
+
+from dotenv import load_dotenv
+
+load_dotenv()
+
+
+class Config:
+    PROJECT_NAME = "AI Fitness Planner"
+    VERSION = "1.0"
+    DATABASE_URL = os.getenv("DATABASE_URL", "")
+    SECRET_KEY = os.getenv("SECRET_KEY", "")
+    ALGORITHM = os.getenv("ALGORITHM", "HS256")
+    ACCESS_TOKEN_EXPIRE_MINUTES = 60
+    GROQ_API_KEY = os.getenv("GROQ_API_KEY", "")
+    GROQ_MODEL = os.getenv("GROQ_MODEL", "llama3-8b-8192")
