+++ backend/app/db/init_db.py
@@ -1,0 +1,6 @@
+from app.db.session import Base, engine
+from app.models import diet, user, workout
+
+
+def init_db():
+    Base.metadata.create_all(bind=engine)
