+++ backend/app/models/__init__.py
@@ -1,0 +1,3 @@
+from app.models.diet import DietPlan
+from app.models.user import User
+from app.models.workout import WorkoutPlan
