--- backend/database.py
@@ -1,33 +1,0 @@
-from config import Config
-from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
-
-DATABASE_URL = Config.DATABASE_URL
-
-# Create database engine
-engine = create_engine(
-    DATABASE_URL,
-    connect_args={"check_same_thread": False} if "sqlite" in DATABASE_URL else {},
-)
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-Base = declarative_base()
-
-
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
-
-
-# Directly create tables on startup
-def initialize_database():
-    from models import DietPlan, User, WorkoutPlan
-
-    Base.metadata.create_all(engine)
-
-
-# Initialize the database
-initialize_database()
