--- backend/models.py
@@ -1,35 +1,0 @@
-from database import Base
-from sqlalchemy import Column, Float, ForeignKey, Integer, String
-from sqlalchemy.orm import relationship
-
-
-class User(Base):
-    __tablename__ = "users"
-
-    id = Column(Integer, primary_key=True, index=True)
-    name = Column(String, nullable=False)
-    email = Column(String, unique=True, nullable=False)
-    hashed_password = Column(String, nullable=False)
-    weight = Column(Float, default=0)
-    height = Column(Float, default=0)
-    goal = Column(String, default="fitness")
-
-
-class WorkoutPlan(Base):
-    __tablename__ = "workout_plans"
-
-    id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"))
-    plan = Column(String, nullable=False)
-
-    user = relationship("User")
-
-
-class DietPlan(Base):
-    __tablename__ = "diet_plans"
-
-    id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"))
-    plan = Column(String, nullable=False)
-
-    user = relationship("User")
